+++ src/Oro/Bundle/AddressBundle/Form/EventListener/AddressCountryAndRegionSubscriber.php
@@ -84,6 +84,9 @@ class AddressCountryAndRegionSubscriber implements EventSubscriberInterface
                 $config['auto_initialize'] = false;
             }
 
+            //@TODO Remove in scope BAP-16504
+            unset($config['cascade_validation']);
+
             $form->add(
                 $this->factory->createNamed(
                     'region',
@@ -114,6 +117,9 @@ class AddressCountryAndRegionSubscriber implements EventSubscriberInterface
             $config = $form->get('region')->getConfig()->getOptions();
             unset($config['choice_list']);
             unset($config['choices']);
+
+            //@TODO Remove in scope BAP-16504
+            unset($config['cascade_validation']);
 
             $config['country'] = $country;
             $config['query_builder'] = $this->getRegionClosure($country);