+++ src/Oro/Bundle/ApiBundle/Processor/MetadataProcessor.php
@@ -5,6 +5,9 @@ namespace Oro\Bundle\ApiBundle\Processor;
 use Oro\Bundle\ApiBundle\Processor\GetMetadata\MetadataContext;
 use Oro\Component\ChainProcessor\ActionProcessor;
 
+/**
+ * The main processor for "get_metadata" action.
+ */
 class MetadataProcessor extends ActionProcessor
 {
     /**