+++ src/Oro/Bundle/SecurityBundle/EventListener/OwnerTreeListener.php
@@ -6,7 +6,7 @@ use Doctrine\ORM\Event\OnFlushEventArgs;
 use Doctrine\ORM\PersistentCollection;
 use Doctrine\ORM\UnitOfWork;
 use Oro\Bundle\SecurityBundle\Owner\OwnerTreeProviderInterface;
-use Symfony\Component\Security\Core\Util\ClassUtils;
+use Symfony\Component\Security\Acl\Util\ClassUtils;
 
 class OwnerTreeListener
 {