+++ src/Oro/Bundle/SegmentBundle/Tests/Unit/Query/DynamicSegmentQueryBuilderTest.php
@@ -44,7 +44,7 @@ class DynamicSegmentQueryBuilderTest extends SegmentDefinitionTestCase
                     []
                 ),
                 new CsrfExtension(
-                    $this->createMock('Symfony\Component\Form\Extension\Csrf\CsrfProvider\CsrfProviderInterface')
+                    $this->createMock('Symfony\Component\Security\Csrf\CsrfTokenManagerInterface')
                 )
                 ]
             )