+++ src/Oro/Bundle/SoapBundle/Controller/Api/Soap/SoapGetController.php
@@ -7,7 +7,7 @@ use Oro\Bundle\SoapBundle\Controller\Api\EntityManagerAwareInterface;
 use Oro\Bundle\SoapBundle\Entity\SoapEntityInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareTrait;
-use Symfony\Component\Security\Core\Util\ClassUtils;
+use Symfony\Component\Security\Acl\Util\ClassUtils;
 
 abstract class SoapGetController implements EntityManagerAwareInterface, SoapApiReadInterface, ContainerAwareInterface
 {